storage: share PropertyTypeDispatcher instead of duplicating it
